+++ test/Spec.hs
@@ -21,6 +21,7 @@ import Test.MockCat.ShouldBeCalledMockMSpec as ShouldBeCalledMockM
 import Test.MockCat.WithMockSpec as WithMock
 import Test.MockCat.ErrorDiffSpec as ErrorDiff
 import Test.MockCat.THCompareSpec as THCompare
+import ReadmeVerifySpec as ReadmeVerify
 import Test.QuickCheck (property)
 import qualified Property.ConcurrentCountProp as ConcurrencyProp
 import qualified Property.LazyEvalProp as LazyEvalProp
@@ -56,6 +57,7 @@ main = hspec $ do
     ShouldBeCalledMockM.spec
     WithMock.spec
     ErrorDiff.spec
+    ReadmeVerify.spec
     describe "Property Concurrency" $ do
       it "total apply count is preserved across threads" $ property ConcurrencyProp.prop_concurrent_total_apply_count
     describe "Property Lazy Evaluation" $ do